+++ src/DrWatson.jl
@@ -51,7 +51,7 @@ end
 
 # Update messages
 const display_update = true
-const update_version = "1.19.0"
+const update_version = "2.0.0"
 const update_name = "update_v$update_version"
 if display_update
 if !isfile(joinpath(@__DIR__, update_name))
@@ -61,6 +61,10 @@ printstyled(stdout,
 
 `savename` no longer replaces `AbstractFloat` values with integer values
 if they two values coincide. I.e. no longer is `1.0` output as `1` in `savename`.
+In this new major release, the following breaking changes have occured:
+1. DrWatson now uses, and suggests using, JLD2.jl instead of BSON.jl
+   for saving files.
+2. The `savename` stuff.
 \n
 """; color = :light_magenta)
 touch(joinpath(@__DIR__, update_name))